perf(PostQueries): improve fetching blogs performance
Ensure there is in memory cache of previous post.
The old posts get returned as the new posts are fetched in the background.
The user gets a response of old posts, while the next request gets the updated in the background.
This is good for performance. A few minutes it takes to fetch old posts is saved.